moved Authentification class to service folder
